+++ provider/provider.go
@@ -79,7 +79,7 @@ func Provider() p.Provider {
 		).
 		WithFunctions(
 			infer.Function(GetGpuTypes{}),
-			infer.Function(GetCpuFlavors{}),
+			infer.Function(GetCPUFlavors{}),
 			infer.Function(GetDataCenters{}),
 		).
 		WithConfig(infer.Config(&Config{})).