+++ sdk/nodejs/types/output.ts
@@ -5,7 +5,7 @@ import * as pulumi from "@pulumi/pulumi";
 import * as inputs from "../types/input";
 import * as outputs from "../types/output";
 
-export interface CpuFlavorOutput {
+export interface CPUFlavorOutput {
     /**
      * Disk limit per vCPU (in GB).
      */